refactor(chat): Group `ChatVaultAPI`, `ListenerColorChat` and `ListenerPrefixChat` into the same package
